4 files fir input and other remarks
