+++ Lab_python_Concurrent_Processing/general-stats.csv
@@ -1,0 +1,9 @@
+Name,Value
+Unknown,23
+Postcard collection,803
+Unknown,90042
+United States of America,203
+Unknown,101956
+Chemical collection,0.003
+Unknown,59033
+Canada,4456
