+++ Lab_python_Concurrent_Processing/object-stats.csv
@@ -1,0 +1,272 @@
+Category,Count
+,229384
+Landmarks,92
+Magnetism,29
+Finishing general,14
+Tools & equipment-trades,1454
+Loose-leaf binding,1
+Studio & set equipment,88
+Dentistry,551
+Lighting device parts,127
+Computer,7
+Clocks,245
+Stitching,4
+Hovercraft parts,1
+Barns & buildings,164
+Typecasting,32
+Industrial vehicles,4
+Test-measurement equipment,364
+Service-test equipment,273
+Lock up/imposition,27
+Surface wind speed & direction measurement,95
+Office equipment,80
+Communications,467
+Navigation instruments &equipment,1
+Photoemissive,47
+Timing devices,23
+Dairying,720
+Cine camera parts,90
+Sewing,14
+Padding & tableting,4
+Radiology,380
+Cutting,18
+Inking,19
+Printing vehicles,3
+Atmospheric pressure measurement,143
+Harvesting,594
+"Precipitation amount, rate &duration measurement",1
+Experimental,23
+Sand glasses,7
+Medical equipment,1041
+Recreational vehicles,20
+Motorcycle parts,21
+Insulators,457
+Visual,11
+Lifting & flying apparatus,21
+Servicing,374
+Miscellaneous equipment,18
+Lighting accessories,15
+Pharmacology,70
+Measurement-position,47
+Sunshine intensity & duration measurement,29
+Food processing,974
+Hydrostatics & fluid mechanics,97
+Light &electromagnetic radiation,1
+Measurement-direction,94
+Tools-Machine,220
+Upper wind speed & direction measurement,55
+Still images,737
+Cycles& cycling,3
+Length,308
+Crop protection,28
+Pneumatics & air pressure,43
+Tooling,121
+Duplication,74
+Merchandising,1304
+Exterior work,38
+Veterinary medicine,58
+Specialised,29
+Patient care & aids for the disabled,208
+Measurement-angles,144
+Telegraphy,766
+Platemaking,114
+Precision tools,38
+Aircraft parts,494
+Demonstration & teaching equipment,476
+Still camera parts,359
+Vessels,70
+Radio,1907
+Mineral exploration,148
+Operations,368
+Motorcycle vehicles,74
+Commemorative,4339
+Folding,5
+Research,1446
+Space science,150
+Tools-Hand,1921
+Environment control,203
+Cloud height & motion measurement,13
+Typesetting,364
+Measurement,720
+Atmospheric humidity measurement,70
+Animal husbandry,162
+Transportation,123
+"Precipitation amount, rate & duration measurement",81
+Observing aids,16
+Candles,3
+Casing-in,10
+Rolling stock,40
+Cottage industries,98
+Thermal & thermal conductivity,38
+Development & maintenance equipment,168
+Tools & equipment,23
+Mass/weight,385
+"Upper air pressure, temperature & humidity measurement",115
+Models,948
+Calendars,23
+Maintenance equipment,79
+Industrial,18
+Make-up,115
+1925,1
+Miscellaneous,4750
+Volume,93
+Atomic &nuclear,7
+Surface wind speed &direction measurement,1
+Tools &equipment-trades,3
+Rounding & backing,5
+Receiving,1139
+Paging & numbering,2
+Still processing,521
+Horticulture,78
+Plumbing,19
+Analog computing devices,78
+Test equipment,22
+Manned spacecraft,209
+Personal gear,162
+Laboratory chemicals,63
+Feeding,1
+Advertising,51
+Memorabilia,13
+Biochemistry,8
+Combination instruments,13
+Telephony,876
+Vacuum tube accessories,113
+Electricity & electronics,274
+Packaging,1
+Petrochemical extraction,18
+Perforating,14
+Motive power,651
+Child care,20
+Chemicals & medications,320
+Airport operations,133
+Digital peripheral devices,943
+Lettering,91
+Furniture,9
+Crop development,1
+Safety equipment,5
+"Example of type of protective glove worn by operators of x-ray eq. lead lined gloves, aprons & face shields were some of the earliest & simplest protective devices used: screening stands and couches, and lead glass viewing windows were also employed. the effect of x-rays on both operators and patients was not well understood for many years, but by 1928 minimum lead values for gloves, etc. had been established.",1
+Lighting devices,429
+Clothing & personal safety apparatus,20
+Regulation &transformation,3
+Acoustics,63
+Aircraft instruments,679
+Atmospheric temperature measurement,233
+Hydrography/oceanography,6
+Moving equipment,1
+Navigation instruments & equipment,937
+Cycles & cycling,1493
+Specific gravity/density,28
+http://source.techno-science.ca/artifacts-artefacts/images/1983.0593.001.aa.cs.png,1
+Motorcycle engineering,21
+Electronics,196
+Automotive engineering,215
+Sundials,3
+Boat & shipbuilding tools & equipment,133
+Laboratory equipment,503
+Light & electromagnetic radiation,848
+Services,1433
+Still cameras,1286
+Civil engineering,40
+Vacuum tube parts,13
+Lamps-lanterns,3
+Conversion,240
+Lumbering,74
+Atomic & nuclear,549
+Digital computing devices,553
+Still camera accessories,818
+Regulation & transformation,92
+Photogrammetry,324
+Fire engines,20
+Transmitting,248
+Fire fighting,275
+Power sources,501
+Image carriers,1273
+Adhesive binding,2
+Naval architecture,34
+Apiculture,179
+Temperature,39
+Early developments,5
+Watches,51
+Hovercraft vehicles,2
+Earth observation,23
+Aircraft,237
+Tool maintenance,127
+Crop processing,104
+Instruments,1061
+Observation equipment,285
+User site,680
+Lamp bulbs,1098
+Cine processing,18
+Petrochemical exploration,24
+Cine cameras,352
+Bundling,1
+Music,340
+Etching/engraving,14
+Public transportation,49
+Measurement-distance,152
+Automotive vehicles,110
+Survey equipment,248
+Mechanics,8
+Ephemera,9
+Protective equipment,1
+Housekeeping,109
+Electromechanical & mechanical peripheral devices,23
+Aviation armament,164
+Geophysics,396
+Alarms,20
+Marbling-edging,5
+Silviculture,125
+Cine miscellaneous,6
+Laundry,418
+Material storage,89
+Animal powered,333
+Telephone,99
+Ruling,26
+Chronometers,9
+Long range forecasting equipment,42
+Television,802
+Image transfer,133
+Extinguishers,59
+Industrial equipment,629
+Military,7
+Upper wind speed &direction measurement,1
+Diving,36
+Vessel parts,129
+Automotive parts,283
+Cultivation,270
+Hunting & trapping,2
+Instrument testing equipment,5
+Poultry equipment,107
+Toilet article,318
+Tool storage,38
+Calculating devices,275
+Drawing & measuring instruments,534
+Expedition artifacts,63
+Evaporation amount measurement,22
+Chemicals &medications,2
+Food service,324
+Generation,257
+Land improvement,127
+Cine camera accessories,155
+Accessories,123
+Archives,4035
+Printed specimens,4355
+Cine images,493
+Analog peripheral devices,2
+Electricity &electronics,3
+Sound,1382
+Human powered,47
+Accessory equipment,87
+Transmission & distribution,54
+Casemaking,2
+Electromechanical & mechanical computing devices,12
+Traffic control,3
+Radar,269
+Cycles &cycling,25
+Power devices,3
+Clothing,2060
+Electricity,157
+Drying,2
+Mineral extraction,12
+Textile working,276
+Crop handling,231
