+++ Lab_python_Concurrent_Processing/tail_recursion.py
@@ -1,0 +1,21 @@
+class Recurse(Exception):
+    def __init__(self, *args, **kwargs):
+        self.args = args
+        self.kwargs = kwargs
+
+
+def recurse(*args, **kwargs):
+    raise Recurse(*args, **kwargs)
+
+
+def tail_recursive(f):
+    def decorated(*args, **kwargs):
+        while True:
+            try:
+                return f(*args, **kwargs)
+            except Recurse as r:
+                args = r.args
+                kwargs = r.kwargs
+                continue
+
+    return decorated
