--- general-stats.csv
@@ -1,9 +1,0 @@
-Name,Value
-Unknown,45
-Postcard collection,803
-Unknown,93895
-United States of America,203
-Unknown,108017
-Chemical collection,0.003
-Unknown,62261
-United States of America,5073
