--- object-stats.csv
@@ -1,273 +1,0 @@
-Category,Count
-,238002
-Lamps-lanterns,3
-Drying,2
-Vessel parts,134
-Calculating devices,276
-Observation equipment,285
-Petrochemical extraction,31
-Transportation,123
-Furniture,9
-Receiving,1139
-Hydrography/oceanography,6
-Casing-in,10
-Atomic &nuclear,7
-Cutting,18
-Civil engineering,40
-Harvesting,595
-Analog computing devices,78
-Combination instruments,13
-Image carriers,1281
-Etching/engraving,14
-Feeding,1
-Upper wind speed & direction measurement,55
-Surface wind speed & direction measurement,98
-Specialised,29
-Public transportation,62
-Crop protection,43
-Alarms,20
-Crop processing,153
-Mineral extraction,42
-Animal powered,335
-Chemicals &medications,2
-http://source.techno-science.ca/artifacts-artefacts/images/1983.0593.001.aa.cs.png,1
-Ephemera,9
-Pneumatics & air pressure,43
-Archives,4144
-Measurement-position,79
-Atmospheric temperature measurement,233
-Child care,60
-Veterinary medicine,58
-Magnetism,29
-Laboratory equipment,520
-Padding & tableting,4
-Service-test equipment,273
-Upper wind speed &direction measurement,1
-Cine processing,18
-Stitching,4
-Patient care & aids for the disabled,221
-Protective equipment,1
-Expedition artifacts,63
-Printed specimens,4356
-Lighting device parts,131
-Temperature,39
-Sewing,14
-"Example of type of protective glove worn by operators of x-ray eq. lead lined gloves, aprons & face shields were some of the earliest & simplest protective devices used: screening stands and couches, and lead glass viewing windows were also employed. the effect of x-rays on both operators and patients was not well understood for many years, but by 1928 minimum lead values for gloves, etc. had been established.",1
-Lumbering,78
-Tools & equipment-trades,4032
-Experimental,23
-Insulators,457
-Measurement-angles,161
-Plumbing,37
-Hovercraft parts,1
-Cottage industries,188
-Sound,1428
-Space science,150
-Operations,380
-Folding,5
-Apiculture,182
-Tooling,121
-Automotive engineering,215
-Lighting accessories,15
-Tool maintenance,127
-Motorcycle parts,21
-Poultry equipment,114
-Typecasting,32
-Commemorative,4416
-Photogrammetry,324
-Sundials,4
-Geophysics,396
-Crop handling,242
-Drawing & measuring instruments,560
-Office equipment,83
-Cycles & cycling,1511
-Food processing,1443
-Industrial vehicles,5
-Digital peripheral devices,993
-Radiology,380
-Fire fighting,278
-Make-up,115
-Lifting & flying apparatus,21
-Cultivation,271
-Visual,13
-Animal husbandry,207
-"Precipitation amount, rate &duration measurement",1
-Electricity,157
-Atmospheric pressure measurement,143
-Survey equipment,274
-Development & maintenance equipment,168
-Digital computing devices,602
-User site,680
-Material storage,89
-Medical equipment,1120
-Demonstration & teaching equipment,476
-Tools & equipment,23
-Watches,75
-Regulation & transformation,93
-Radar,269
-Accessories,136
-Measurement-distance,153
-Biochemistry,100
-Chronometers,9
-Services,1433
-Computer,7
-Electronics,196
-"Upper air pressure, temperature & humidity measurement",115
-Tools-Hand,2322
-Cine cameras,364
-Test equipment,22
-Printing vehicles,3
-Tool storage,39
-Models,1090
-Mass/weight,385
-,1
-Crop development,1
-Still cameras,1319
-Mineral exploration,225
-Measurement-direction,95
-Telephony,1002
-Generation,309
-Rolling stock,40
-Specific gravity/density,28
-Aircraft instruments,679
-Vacuum tube accessories,113
-Bundling,1
-Adhesive binding,2
-Measurement,720
-Memorabilia,15
-Chemicals & medications,673
-Instrument testing equipment,5
-Toilet article,350
-Land improvement,146
-Automotive parts,288
-Ruling,26
-Human powered,63
-Calendars,23
-Dentistry,552
-Recreational vehicles,33
-Clothing & personal safety apparatus,38
-Length,312
-Housekeeping,150
-Television,890
-Photoemissive,47
-Test-measurement equipment,374
-Landmarks,92
-Navigation instruments & equipment,984
-Lettering,91
-Lock up/imposition,27
-Automotive vehicles,111
-Telephone,99
-Motorcycle vehicles,75
-Exterior work,39
-Atomic & nuclear,552
-Electricity & electronics,274
-Barns & buildings,176
-Music,348
-Motorcycle engineering,21
-Merchandising,1395
-Hovercraft vehicles,2
-Perforating,14
-Cloud height & motion measurement,13
-Instruments,1433
-Electricity &electronics,3
-Navigation instruments &equipment,1
-Cycles& cycling,3
-Evaporation amount measurement,22
-Inking,19
-Atmospheric humidity measurement,70
-Manned spacecraft,209
-Vessels,70
-Long range forecasting equipment,42
-Industrial,18
-Clocks,246
-Still processing,532
-Platemaking,114
-Power sources,534
-Power devices,3
-Research,1640
-Cine camera accessories,178
-Accessory equipment,87
-Electromechanical & mechanical peripheral devices,23
-Thermal & thermal conductivity,38
-Studio & set equipment,88
-Telegraphy,766
-Military,7
-Cine camera parts,93
-Cycles &cycling,25
-Timing devices,23
-Precision tools,40
-Casemaking,2
-Marbling-edging,5
-Horticulture,215
-Analog peripheral devices,2
-Laboratory chemicals,74
-Pharmacology,70
-Aircraft,264
-Paging & numbering,2
-Transmitting,248
-Hunting & trapping,2
-Motive power,652
-Extinguishers,59
-Rounding & backing,5
-Light &electromagnetic radiation,1
-Mechanics,8
-Conversion,245
-Finishing general,14
-Silviculture,125
-Radio,2055
-Candles,3
-Airport operations,133
-Advertising,51
-Image transfer,155
-"Precipitation amount, rate & duration measurement",132
-Traffic control,3
-Servicing,375
-Loose-leaf binding,1
-Still camera accessories,857
-Environment control,205
-Sunshine intensity & duration measurement,29
-Aviation armament,164
-Observing aids,16
-Miscellaneous,5346
-Tools-Machine,221
-Hydrostatics & fluid mechanics,97
-Lighting devices,501
-Maintenance equipment,79
-Communications,495
-Typesetting,370
-Safety equipment,5
-Sand glasses,7
-Cine images,496
-Early developments,5
-Regulation &transformation,3
-Personal gear,338
-Still images,924
-Diving,37
-Volume,96
-Light & electromagnetic radiation,906
-Surface wind speed &direction measurement,1
-Naval architecture,40
-Food service,348
-Industrial equipment,645
-Moving equipment,1
-1925,1
-Dairying,746
-Duplication,74
-Tools &equipment-trades,3
-Lamp bulbs,1106
-Miscellaneous equipment,20
-Petrochemical exploration,24
-Clothing,2092
-Still camera parts,366
-Vacuum tube parts,13
-Transmission & distribution,55
-Fire engines,20
-Packaging,2
-Acoustics,63
-Boat & shipbuilding tools & equipment,134
-Textile working,314
-Aircraft parts,507
-Earth observation,24
-Electromechanical & mechanical computing devices,12
-Laundry,418
-Cine miscellaneous,6
